requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
djoser
djangorestframework_simplejwt
django-rest-passwordreset
ujson
xlrd
rest_condition